config: Change interval from u8 to u16 to allow larger values

+++ src/config.rs
@@ -11,7 +11,7 @@ use snafu::ResultExt;
 use crate::error::*;
 
 #[derive(Clone, Copy, Debug, Deserialize)]
-pub struct Interval(pub u8);
+pub struct Interval(pub u16);
 
 impl Interval {
     pub fn to_duration(self) -> Duration {
